Rename duke package to junjie package
