+++ src/main/java/junjie/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package junjie;
 
 /**
  * The Ui class handles the user interface of the application.
