+++ src/main/java/junjie/commands/DefaultCommand.java
@@ -1,7 +1,7 @@
-package duke.commands;
+package junjie.commands;
 
-import duke.TaskList;
-import duke.Ui;
+import junjie.TaskList;
+import junjie.Ui;
 
 /**
  * Represents a command that is not recognised by the chat bot.
