+++ src/main/java/junjie/commands/ExitCommand.java
@@ -1,7 +1,7 @@
-package duke.commands;
+package junjie.commands;
 
-import duke.TaskList;
-import duke.Ui;
+import junjie.TaskList;
+import junjie.Ui;
 
 /**
  * Represents a command to exit the program.
