+++ src/main/java/junjie/commands/MarkCommand.java
@@ -1,8 +1,8 @@
-package duke.commands;
+package junjie.commands;
 
-import duke.TaskList;
-import duke.Ui;
-import duke.tasks.Task;
+import junjie.TaskList;
+import junjie.Ui;
+import junjie.tasks.Task;
 
 /**
  * A command to mark a task as done.
