+++ src/main/java/junjie/commands/TodoCommand.java
@@ -1,10 +1,10 @@
-package duke.commands;
+package junjie.commands;
 
-import duke.TaskList;
-import duke.Ui;
-import duke.exceptions.InvalidArgumentException;
-import duke.tasks.Task;
-import duke.tasks.TodoTask;
+import junjie.TaskList;
+import junjie.Ui;
+import junjie.exceptions.InvalidArgumentException;
+import junjie.tasks.Task;
+import junjie.tasks.TodoTask;
 
 /**
  * Represents a command to add a todo task to the task list.
